requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
supabase==2.3.4
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0

